Moving Towards release 1.2
